Squashed 'h5hut/' changes from 7e7a85f14..3a9e6d833
3a9e6d833 Merge branch 'int8' into 'master'
9d4c88443 Add support for hdf5 int8 types (int8, uint8)
40644d0c9 example from Eurohack24 added
c41bd6ec5 Merge branch '21-hdf5-1-12-and-newer-support' into 'master'
a69a88c49 version bumped to 2.0.0rc7
4088d02f5 hdf5 >=1.12.0 support added
5f57c2349 Merge branch '20-unknown-macro-ac_msg_erro' into 'master'
b71fa8c9d LT_INIT added to configure.ac, should fix unknown AC_MSG_ERROR

git-subtree-dir: h5hut
git-subtree-split: 3a9e6d8335e13a154cddc64cf1589b2adbe85e9a

+++ h5hut/src/h5core/private/h5_attribs.h
@@ -99,7 +99,7 @@ h5priv_write_attrib (
 	                   attrib_name,
 	                   (long long int)attrib_type,
 	                   attrib_value,
-	                   attrib_nelem);
+	                   (long long unsigned)attrib_nelem);
 	hid_t space_id;
 	hid_t attrib_id;
 	hid_t hdf5_type;
@@ -147,7 +147,7 @@ h5priv_append_attrib (
 	                   attrib_name,
 	                   (long long int)attrib_type,
 	                   attrib_value,
-	                   attrib_nelem);
+	                   (long long unsigned)attrib_nelem);
 	h5_err_t exists;
 	TRY (exists = hdf5_attribute_exists (id, attrib_name));
 	if (exists) {
